requirements.txt
Flask==3.1.1
discord.py==2.3.2